requirements.txt
torch 
torchvision 
torchaudio
bertopic
logparser3
pyarrow
fastparquet
networkx
plotly
dash
ipykernel
matplotlib
pandas
numpy
pcalg
torch_geometric
pydot
tqdm
graphviz
pygraphviz